backend/requirements.txt
# Core framework
fastapi==0.115.2
uvicorn[standard]==0.32.0

# Banco de dados
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validações e configurações
pydantic[email]==2.9.2
pydantic-settings==2.4.0

# Uploads e formulários
python-multipart==0.0.9

# Utilitários úteis em dev
python-dotenv==1.0.1
